wip: adding introspect guard to be able to use introspection to log users in
